feat: Add value attribute to input field
This commit adds a new `value` attribute to the input field in the `field.blade.php` file. The value is set using the `$value` variable. This change allows for pre-populating the input field with a default value.

Additionally, in the `LivewireGooglePlacesAutocomplete.php` file, a new public property `$value` is added. This property can be used to store and retrieve the current value of the input field.

These changes enhance the functionality of the autocomplete feature by providing support for setting and retrieving values in the input field.